move product fixture to a new module
